backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
sqlalchemy>=2.0.30
psycopg2-binary>=2.9.9
requests>=2.32.0
python-dotenv>=1.0.1
